Move ssl attributes to SslConfiguration and remove deprecated method
